pyproject.toml
[project]
name = "multi-tenant-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic==1.17.2",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.0",
    "argon2-cffi>=25.1.0",
    "asyncpg==0.31.0",
    "click==8.3.1",
    "colorama==0.4.6",
    "email-validator>=2.3.0",
    "fastapi==0.128.0",
    "greenlet==3.3.0",
    "h11==0.16.0",
    "httptools==0.7.1",
    "idna==3.11",
    "mako==1.3.10",
    "markupsafe==3.0.3",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.12.0",
    "pyjwt[crypto]>=2.10.1",
    "python-dotenv==1.2.1",
    "pyyaml==6.0.3",
    "redis==7.1.0",
    "sqlalchemy==2.0.45",
    "starlette==0.50.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "uuid-utils>=0.12.0",
    "uvicorn==0.40.0",
    "watchfiles==1.1.1",
    "websockets==15.0.1",
]
